Put event-centric files into an Events directory for more clarity
