+++ src/Components/Web.JS/src/Rendering/Events/EventDispatcher.ts
@@ -1,4 +1,12 @@
-import { EventDescriptor } from './BrowserRenderer';
+import { EventFieldInfo } from './EventFieldInfo';
+
+export interface EventDescriptor {
+  browserRendererId: number;
+  eventHandlerId: number;
+  eventName: string;
+  eventFieldInfo: EventFieldInfo | null;
+}
+
 type EventDispatcher = (eventDescriptor: EventDescriptor, eventArgs: any) => void;
 
 let eventDispatcherInstance: EventDispatcher;
